+++ CHANGES.rst
@@ -4,6 +4,10 @@ Changelog
 1.0a3 (unreleased)
 ^^^^^^^^^^^^^^^^^^
 
+  * A visão padrão de Agenda agora exibe a AgendaDiaria
+    do dia se estiver publicada ou a mais recente.
+    [ericof]
+
   * Título de AgendaDiaria fica no formato
     "Agenda de <autoridade> para <data>"
     [ericof]